+++ rust/cubesql/cubesql/src/compile/engine/df/wrapper.rs
@@ -32,8 +32,16 @@ use itertools::Itertools;
 use regex::{Captures, Regex};
 use serde::{Deserialize, Serialize};
 use std::{
-    any::Any, cmp::min, collections::HashMap, convert::TryInto, fmt, future::Future, iter,
-    pin::Pin, result, sync::Arc,
+    any::Any,
+    cmp::min,
+    collections::HashMap,
+    convert::TryInto,
+    fmt,
+    future::Future,
+    iter,
+    pin::Pin,
+    result,
+    sync::{Arc, LazyLock},
 };
 
 #[derive(Debug, Clone, Deserialize)]
@@ -160,12 +168,12 @@ impl SqlQuery {
     }
 
     pub fn finalize_query(&mut self, sql_templates: Arc<SqlTemplates>) -> Result<()> {
+        static REGEX: LazyLock<Regex> = LazyLock::new(|| Regex::new(r"\$(\d+)\$").unwrap());
+
         let mut params = Vec::new();
         let mut rendered_params = HashMap::new();
-        let regex = Regex::new(r"\$(\d+)\$")
-            .map_err(|e| DataFusionError::Execution(format!("Can't parse regex: {}", e)))?;
         let mut res = Ok(());
-        let replaced_sql = regex.replace_all(self.sql.as_str(), |c: &Captures<'_>| {
+        let replaced_sql = REGEX.replace_all(self.sql.as_str(), |c: &Captures<'_>| {
             let param = c.get(1).map(|x| x.as_str());
             match self.render_param(sql_templates.clone(), param, &rendered_params, params.len()) {
                 Ok((param_index, param, push_param)) => {
@@ -260,9 +268,7 @@ pub struct SqlGenerationResult {
     pub request: TransportLoadRequestQuery,
 }
 
-lazy_static! {
-    static ref DATE_PART_REGEX: Regex = Regex::new("^[A-Za-z_ ]+$").unwrap();
-}
+static DATE_PART_REGEX: LazyLock<Regex> = LazyLock::new(|| Regex::new("^[A-Za-z_ ]+$").unwrap());
 
 macro_rules! generate_sql_for_timestamp {
     (@generic $value:ident, $value_block:expr, $sql_generator:expr, $sql_query:expr) => {
@@ -950,8 +956,9 @@ impl CubeScanWrapperNode {
         ungrouped_scan_node: Option<Arc<CubeScanNode>>,
         subqueries: Arc<HashMap<String, String>>,
     ) -> result::Result<(Vec<AliasedColumn>, SqlQuery), CubeError> {
-        let non_id_regex = Regex::new(r"[^a-zA-Z0-9_]")
-            .map_err(|e| CubeError::internal(format!("Can't parse regex: {}", e)))?;
+        static NON_ID_REGEX: LazyLock<Regex> =
+            LazyLock::new(|| Regex::new(r"[^a-zA-Z0-9_]").unwrap());
+
         let mut aliased_columns = Vec::new();
         for original_expr in exprs {
             let expr = if let Some(column_remapping) = column_remapping.as_ref() {
@@ -1001,7 +1008,7 @@ impl CubeScanWrapperNode {
 
             let alias = if can_rename_columns {
                 let alias = expr_name(&expr, &schema)?;
-                let mut truncated_alias = non_id_regex
+                let mut truncated_alias = NON_ID_REGEX
                     .replace_all(&alias, "_")
                     .trim_start_matches("_")
                     .to_lowercase();